// __generated__/api-react-default-fetch-unwrap-response-data/api.ts
export interface TypeApiResponse {
  /** @format int32 */
  code?: number;
  type?: string;
  message?: string;
}

export interface TypeCategory {
  /** @format int64 */
  id?: number;
  name?: string;
}

export interface TypePet {
  /** @format int64 */
  id?: number;
  category?: TypeCategory;
  /** @example "doggie" */
  name: string;
  photoUrls: string[];
  tags?: TypeTag[];
  /** pet status in the store */
  status?: "available" | "pending" | "sold";
}

export interface TypeTag {
  /** @format int64 */
  id?: number;
  name?: string;
}

export interface TypeOrder {
  /** @format int64 */
  id?: number;
  /** @format int64 */
  petId?: number;
  /** @format int32 */
  quantity?: number;
  /** @format date-time */
  shipDate?: string;
  /** Order Status */
  status?: "placed" | "approved" | "delivered";
  complete?: boolean;
}

export interface TypeUser {
  /** @format int64 */
  id?: number;
  username?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  password?: string;
  phone?: string;
  /**
   * User Status
   * @format int32
   */
  userStatus?: number;
}

export type QueryParamsType = Record<string | number, any>;
export type ResponseFormat = keyof Omit<Body, "body" | "bodyUsed">;

export interface FullRequestParams extends Omit<RequestInit, "body"> {
  /** set parameter to `true` for call `securityWorker` for this request */
  secure?: boolean;
  /** request path */
  path: string;
  /** content type of request body */
  type?: ContentType;
  /** query params */
  query?: QueryParamsType;
  /** format of response (i.e. response.json() -> format: "json") */
  format?: ResponseFormat;
  /** request body */
  body?: unknown;
  /** base url */
  baseUrl?: string;
  /** request cancellation token */
  cancelToken?: CancelToken;
}

export type RequestParams = Omit<FullRequestParams, "body" | "method" | "query" | "path">;
// ...
type CancelToken = Symbol | string | number;

export enum ContentType {
  Json = "application/json",
  FormData = "multipart/form-data",
  UrlEncoded = "application/x-www-form-urlencoded",
  Text = "text/plain",
}